A mini study on smoking propension in the USA
What are the causes for young adults to become regular smokers? what variables are involved? This is mini analysis on possible correlations and on what variables have the largest effects to produce regular smoking habits.
This analysis is based on a original study and data provided for academic purposes by Dr. Lisa Dierker
https://doi.org/10.1016/j.drugalcdep.2012.08.026